@@ -10,34 +10,34 @@
     "permissions": "chown -R $USER:$USER bootstrap/cache storage public/uploads"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0-beta.40",
-    "@babel/polyfill": "^7.0.0-beta.40",
-    "@babel/preset-env": "^7.0.0-beta.40",
-    "autoprefixer": "^8.1.0",
-    "babel-loader": "^8.0.0-beta.0",
-    "css-loader": "^0.28.10",
+    "@babel/core": "^7.1.5",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.1.5",
+    "autoprefixer": "^8.6.5",
+    "babel-loader": "^8.0.4",
+    "css-loader": "^0.28.11",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "livereload": "^0.7.0",
-    "node-sass": "^4.9.2",
-    "npm-run-all": "^4.1.2",
-    "postcss-loader": "^2.1.1",
-    "sass-loader": "^7.0.1",
+    "node-sass": "^4.10.0",
+    "npm-run-all": "^4.1.3",
+    "postcss-loader": "^2.1.6",
+    "sass-loader": "^7.1.0",
     "style-loader": "^0.21.0",
-    "uglifyjs-webpack-plugin": "^1.2.3",
-    "webpack": "^4.16.3",
-    "webpack-cli": "^2.0.11"
+    "uglifyjs-webpack-plugin": "^1.3.0",
+    "webpack": "^4.25.1",
+    "webpack-cli": "^3.1.2"
   },
   "dependencies": {
     "axios": "^0.18.0",
-    "clipboard": "^2.0.0",
-    "codemirror": "^5.26.0",
-    "dropzone": "^5.4.0",
+    "clipboard": "^2.0.1",
+    "codemirror": "^5.41.0",
+    "dropzone": "^5.5.1",
     "jquery": "^3.3.1",
     "jquery-sortable": "^0.9.13",
-    "markdown-it": "^8.3.1",
-    "markdown-it-task-lists": "^2.0.0",
-    "vue": "^2.2.6",
-    "vuedraggable": "^2.14.1"
+    "markdown-it": "^8.4.2",
+    "markdown-it-task-lists": "^2.1.1",
+    "vue": "^2.5.17",
+    "vuedraggable": "^2.16.0"
   },
   "browser": {
     "vue": "vue/dist/vue.common.js"
