@@ -52,17 +52,51 @@ return [
     'search_content_type' => '种类',
     'search_exact_matches' => '精确匹配',
     'search_tags' => '标签搜索',
+    'search_options' => '选项',
     'search_viewed_by_me' => '我看过的',
     'search_not_viewed_by_me' => '我没看过的',
     'search_permissions_set' => '权限设置',
     'search_created_by_me' => '我创建的',
     'search_updated_by_me' => '我更新的',
+    'search_date_options' => '日期选项',
     'search_updated_before' => '在此之前更新',
     'search_updated_after' => '在此之后更新',
     'search_created_before' => '在此之前创建',
     'search_created_after' => '在此之后创建',
     'search_set_date' => '设置日期',
     'search_update' => '只显示更新操作',
+
+    /**
+     * Shelves
+     */
+    'shelf' => '书架',
+    'shelves' => '书架',
+    'shelves_long' => '书架',
+    'shelves_empty' => '当前未创建书架',
+    'shelves_create' => '创建新书架',
+    'shelves_popular' => '热门书架',
+    'shelves_new' => '新书架',
+    'shelves_popular_empty' => '最热门的书架',
+    'shelves_new_empty' => '最新创建的书架',
+    'shelves_save' => '保存书架',
+    'shelves_books' => '书籍已在此书架里',
+    'shelves_add_books' => '将书籍加入此书架',
+    'shelves_drag_books' => '拖动图书将其添加到此书架',
+    'shelves_empty_contents' => '这个书架没有分配图书',
+    'shelves_edit_and_assign' => '编辑书架以分配图书',
+    'shelves_edit_named' => '编辑书架 :name',
+    'shelves_edit' => '编辑书架',
+    'shelves_delete' => '删除书架',
+    'shelves_delete_named' => '删除书架 :name',
+    'shelves_delete_explain' => "此操作将删除书架 ':name'。 其中包含的图书不会被删除。",
+    'shelves_delete_confirmation' => '您确定要删除此书架吗？',
+    'shelves_permissions' => '书架权限',
+    'shelves_permissions_updated' => '书架权限已更新',
+    'shelves_permissions_active' => '书架权限激活',
+    'shelves_copy_permissions_to_books' => '将权限复制到图书',
+    'shelves_copy_permissions' => '复制权限',
+    'shelves_copy_permissions_explain' => '这会将此书架的当前权限设置应用于其中包含的所有图书。 在激活之前，请确保已保存对此书架权限的任何更改。',
+    'shelves_copy_permission_success' => '书架权限复制到图书 :count ',
 
     /**
      * Books
@@ -162,9 +196,13 @@ return [
     'pages_md_preview' => '预览',
     'pages_md_insert_image' => '插入图片',
     'pages_md_insert_link' => '插入实体链接',
+	'pages_md_insert_drawing' => '插入图表',
     'pages_not_in_chapter' => '本页面不在某章节中',
     'pages_move' => '移动页面',
     'pages_move_success' => '页面已移动到「:parentName」',
+	'pages_copy' => '复制页面',
+    'pages_copy_desination' => '复制目的地',
+    'pages_copy_success' => '页面复制完成',
     'pages_permissions' => '页面权限',
     'pages_permissions_success' => '页面权限已更新',
     'pages_revision' => '修订',
@@ -195,13 +233,17 @@ return [
         'message' => ':time，:start。注意不要覆盖对方的更新！',
     ],
     'pages_draft_discarded' => '草稿已丢弃，编辑器已更新到当前页面内容。',
+	'pages_specific' => '具体页面',
 
     /**
      * Editor sidebar
      */
     'page_tags' => '页面标签',
+	'chapter_tags' => '章节标签',
+	'book_tags' => '图书标签',
+    'shelf_tags' => '书架标签',
     'tag' => '标签',
-    'tags' =>  '',
+    'tags' =>  '标签',
     'tag_value' => '标签值 (Optional)',
     'tags_explain' => "添加一些标签以更好地对您的内容进行分类。\n您可以为标签分配一个值，以进行更深入的组织。",
     'tags_add' => '添加另一个标签',
@@ -245,6 +287,7 @@ return [
      */
     'comment' => '评论',
     'comments' => '评论',
+	'comment_add' => '添加评论',
     'comment_placeholder' => '在这里评论',
     'comment_count' => '{0} 无评论|[1,*] :count条评论',
     'comment_save' => '保存评论',
