@@ -34,11 +34,14 @@ return [
     'path_not_writable' => '无法上传到文件路径“:filePath”，请确保它可写入服务器。',
     'cannot_get_image_from_url' => '无法从 :url 中获取图片',
     'cannot_create_thumbs' => '服务器无法创建缩略图，请检查您是否安装了GD PHP扩展。',
-    'server_upload_limit' => '上传图片时发生错误。',
+    'server_upload_limit' => '服务器不允许上传此大小的文件。 请尝试较小的文件。',
+	'uploaded'  => 'The server does not allow uploads of this size. Please try a smaller file size.',
     'image_upload_error' => '上传图片时发生错误',
+	'image_upload_type_error' => '上传的图像类型无效',
 
     // Attachments
     'attachment_page_mismatch' => '附件更新期间的页面不匹配',
+	'attachment_not_found' => '找不到附件',
 
     // Pages
     'page_draft_autosave_fail' => '无法保存草稿，确保您在保存页面之前已经连接到互联网',
@@ -46,6 +49,7 @@ return [
 
     // Entities
     'entity_not_found' => '未找到实体',
+	'bookshelf_not_found' => '未找到书架',
     'book_not_found' => '未找到图书',
     'page_not_found' => '未找到页面',
     'chapter_not_found' => '未找到章节',
