@@ -81,6 +81,7 @@ return [
     'role_details' => '角色详细信息',
     'role_name' => '角色名',
     'role_desc' => '角色简述',
+	'role_external_auth_id' => '外部身份认证ID',
     'role_system' => '系统权限',
     'role_manage_users' => '管理用户',
     'role_manage_roles' => '管理角色与角色权限',
@@ -89,6 +90,7 @@ return [
     'role_manage_settings' => '管理App设置',
     'role_asset' => '资源许可',
     'role_asset_desc' => '对系统内资源的默认访问许可将由这些权限控制。单独设置在书籍，章节和页面上的权限将覆盖这里的权限设定。',
+	'role_asset_admins' => '管理员可自动获得对所有内容的访问权限，但这些选项可能会显示或隐藏UI选项。',
     'role_all' => '全部的',
     'role_own' => '拥有的',
     'role_controlled_by_asset' => '由其所在的资源来控制',
@@ -109,7 +111,6 @@ return [
     'users_external_auth_id' => '外部身份认证ID',
     'users_password_warning' => '如果您想更改密码，请填写以下内容：',
     'users_system_public' => '此用户代表访问您的App的任何访客。它不能用于登录，而是自动分配。',
-    'users_books_view_type' => '图书浏览布局偏好',
     'users_delete' => '删除用户',
     'users_delete_named' => '删除用户 :userName',
     'users_delete_warning' => '这将从系统中完全删除名为 \':userName\' 的用户。',
