@@ -52,17 +52,51 @@ return [
     'search_content_type' => '種類',
     'search_exact_matches' => '精確符合',
     'search_tags' => '標籤搜尋',
+    'search_options' => '選項',
     'search_viewed_by_me' => '我看過的',
     'search_not_viewed_by_me' => '我沒看過的',
     'search_permissions_set' => '權限設定',
     'search_created_by_me' => '我建立的',
     'search_updated_by_me' => '我更新的',
+    'search_date_options' => '日期選項',
     'search_updated_before' => '在此之前更新',
     'search_updated_after' => '在此之後更新',
     'search_created_before' => '在此之前建立',
     'search_created_after' => '在此之後建立',
     'search_set_date' => '設定日期',
     'search_update' => '更新搜尋結果',
+
+    /**
+     * Shelves
+     */
+    'shelf' => '書架',
+    'shelves' => '書架',
+    'shelves_long' => '書架',
+    'shelves_empty' => '不存在已建立的書架',
+    'shelves_create' => '建立書架',
+    'shelves_popular' => '熱門書架',
+    'shelves_new' => '新書架',
+    'shelves_popular_empty' => '最受歡迎的書架將出現在這裡。',
+    'shelves_new_empty' => '最近建立的書架將出現在這裡。',
+    'shelves_save' => '儲存書架',
+    'shelves_books' => '此書架上的書本',
+    'shelves_add_books' => '將書本添加到此書架中',
+    'shelves_drag_books' => '拖動書本到此處來將它添加至此書架中',
+    'shelves_empty_contents' => '此書架沒有分配任何書本',
+    'shelves_edit_and_assign' => '編輯書架以分配書本',
+    'shelves_edit_named' => '編輯書架「:name」',
+    'shelves_edit' => '編輯書架',
+    'shelves_delete' => '刪除書架',
+    'shelves_delete_named' => '刪除書架「:name」',
+    'shelves_delete_explain' => "這將刪除名為「:name」的書架。包含在其中的書本不會被刪除。",
+    'shelves_delete_confirmation' => '您確定要刪除此書架嗎？',
+    'shelves_permissions' => '書架權限',
+    'shelves_permissions_updated' => '書架權限已更新',
+    'shelves_permissions_active' => '已啟用此書架的自訂權限',
+    'shelves_copy_permissions_to_books' => '將權限複製到書本',
+    'shelves_copy_permissions' => '複製權限',
+    'shelves_copy_permissions_explain' => '這會將此書架目前的權限設定套用到所有包含的書本上。在生效之前，請確認您已儲存任何對此書架權限的變更。',
+    'shelves_copy_permission_success' => '已將書架的權限複製到:count本書上',
 
     /**
      * Books
@@ -92,7 +126,7 @@ return [
     'books_empty_or' => '或',
     'books_empty_sort_current_book' => '排序目前書本',
     'books_empty_add_chapter' => '加入章節',
-    'books_permissions_active' => '有效的書本權限',
+    'books_permissions_active' => '已啟用此書本的自訂權限',
     'books_search_this' => '搜尋這本書',
     'books_navigation' => '書本導覽',
     'books_sort' => '排序書本內容',
@@ -121,7 +155,7 @@ return [
     'chapter_move_success' => '章節移動到「:bookName」',
     'chapters_permissions' => '章節權限',
     'chapters_empty' => '本章目前沒有頁面。',
-    'chapters_permissions_active' => '有效的章節權限',
+    'chapters_permissions_active' => '已啟用此章節的自訂權限',
     'chapters_permissions_success' => '章節權限已更新',
     'chapters_search_this' => '從本章節搜尋',
 
@@ -166,6 +200,9 @@ return [
     'pages_not_in_chapter' => '本頁面不在某章節中',
     'pages_move' => '移動頁面',
     'pages_move_success' => '頁面已移動到「:parentName」',
+    'pages_copy' => '複製頁面',
+    'pages_copy_desination' => '複製的目的地',
+    'pages_copy_success' => '頁面已成功複製',
     'pages_permissions' => '頁面權限',
     'pages_permissions_success' => '頁面權限已更新',
     'pages_revision' => '修訂',
@@ -183,7 +220,7 @@ return [
     'pages_revisions_none' => '此頁面沒有修訂',
     'pages_copy_link' => '複製連結',
     'pages_edit_content_link' => '编辑内容',
-    'pages_permissions_active' => '有效的頁面權限',
+    'pages_permissions_active' => '已啟用此頁面的自訂權限',
     'pages_initial_revision' => '初次發布',
     'pages_initial_name' => '新頁面',
     'pages_editing_draft_notification' => '您正在編輯在 :timeDiff 內儲存的草稿.',
@@ -196,11 +233,15 @@ return [
         'message' => ':time，:start。注意不要覆蓋到對方的更新。',
     ],
     'pages_draft_discarded' => '草稿已丟棄，編輯器已更新到目前頁面內容。',
+    'pages_specific' => '指定頁面',
 
     /**
      * Editor sidebar
      */
     'page_tags' => '頁面標籤',
+    'chapter_tags' => '章節標籤',
+    'book_tags' => '書本標籤',
+    'shelf_tags' => '書架標籤',
     'tag' => '標籤',
     'tags' =>  '',
     'tag_value' => '標籤值 (非必要)',
@@ -246,6 +287,7 @@ return [
      */
     'comment' => '評論',
     'comments' => '評論',
+    'comment_add' => '新增評論',
     'comment_placeholder' => '在這裡評論',
     'comment_count' => '{0} 無評論|[1,*] :count條評論',
     'comment_save' => '儲存評論',
